jd-collector-agent/requirements.txt
playwright
python-dotenv
google-genai
